moved api processing to a separate file
