+++ src/App.tsx
@@ -1,24 +1,8 @@
 import React from 'react';
 import './App.scss';
 
-import todosFromServer from './api/todos';
-import usersFromServer from './api/users';
-
-import { User } from './types/User';
-import { Todo } from './types/Todo';
 import { TodoList } from './components/TodoList';
-
-function getUser(userId: number): User | null {
-  const foundUser = usersFromServer.find(user => user.id === userId);
-
-  // if there is no user with a given userId
-  return foundUser || null;
-}
-
-export const todos: Todo[] = todosFromServer.map(todo => ({
-  ...todo,
-  user: getUser(todo.userId),
-}));
+import { todos } from './prepareApiData';
 
 export const App: React.FC = () => (
   <div className="App">
